+++ CheckboxItem/CheckboxItem.js
@@ -82,6 +82,34 @@ const CheckboxItemBase = kind({
 		icon: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
 
 		/**
+		 * Enables the "indetermine" state.
+		 *
+		 * An indeterminate, mixed, or half-selected state is typically used in a hierarchy or group
+		 * to represent that some, not all, children are selected.
+		 *
+		 * @type {Boolean}
+		 * @default false
+		 * @public
+		 */
+		indeterminate: PropTypes.bool,
+
+		/**
+		 * Sets an icon to be used in the `indeterminate` state
+		 *
+		 * May be specified as either:
+		 *
+		 * * A string that represents an icon from the [iconList]{@link ui/Icon.Icon.iconList},
+		 * * An HTML entity string, Unicode reference or hex value (in the form '0x...'),
+		 * * A URL specifying path to an icon image, or
+		 * * An object representing a resolution independent resource (See {@link ui/resolution})
+		 *
+		 * @type {String}
+		 * @default 'minus'
+		 * @public
+		 */
+		indeterminateIcon: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+
+		/**
 		 * If true the checkbox will be selected.
 		 *
 		 * @type {Boolean}
@@ -101,7 +129,7 @@ const CheckboxItemBase = kind({
 		publicClassNames: ['checkboxItem']
 	},
 
-	render: ({children, css, icon, selected, ...rest}) => (
+	render: ({children, css, icon, indeterminate, indeterminateIcon, selected, ...rest}) => (
 		<Item
 			data-webos-voice-intent="SelectCheckItem"
 			role="checkbox"
@@ -109,7 +137,15 @@ const CheckboxItemBase = kind({
 			selected={selected}
 			css={css}
 		>
-			<CheckboxBase selected={selected} slot="slotBefore" css={css}>{icon}</CheckboxBase>
+			<CheckboxBase
+				selected={selected}
+				indeterminate={indeterminate}
+				indeterminateIcon={indeterminateIcon}
+				slot="slotBefore"
+				css={css}
+			>
+				{icon}
+			</CheckboxBase>
 			{children}
 		</Item>
 	)