+++ samples/sampler/stories/default/FormCheckboxItem.js
@@ -4,13 +4,14 @@ import {mergeComponentMetadata} from '@enact/storybook-utils';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 
+import Checkbox, {CheckboxBase} from '@enact/sandstone/Checkbox';
 import FormCheckboxItem, {FormCheckboxItemBase} from '@enact/sandstone/FormCheckboxItem';
 import Item, {ItemBase} from '@enact/sandstone/Item';
 
 import iconNames from './icons';
 
 FormCheckboxItem.displayName = 'FormCheckboxItem';
-const Config = mergeComponentMetadata('FormCheckboxItem', ItemBase, Item, FormCheckboxItemBase, FormCheckboxItem);
+const Config = mergeComponentMetadata('FormCheckboxItem', ItemBase, Item, CheckboxBase, Checkbox, FormCheckboxItemBase, FormCheckboxItem);
 
 storiesOf('Sandstone', module)
 	.add(
@@ -20,6 +21,8 @@ storiesOf('Sandstone', module)
 				<FormCheckboxItem
 					disabled={boolean('disabled', Config)}
 					icon={select('icon', ['', ...iconNames], Config)}
+					indeterminate={boolean('indeterminate', Config)}
+					indeterminateIcon={select('indeterminateIcon', ['', ...iconNames], Config)}
 					inline={boolean('inline', Config)}
 					label={text('label', Config, '')}
 					labelPosition={select('labelPosition', ['', 'above', 'after', 'before', 'below'], Config, '')}